==> 03_Feature_engineering.Rmd <==
---
title: "03 Feature engineering"
author: "Dominik Klepl"
date: "11/15/2018"
output: html_document
---

## Load data
```{r}
data_c = read.csv("clean_data/ML_ready/coordination_data.csv")
data_i = read.csv("clean_data/ML_ready/interviewer_data.csv")
data_p = read.csv("clean_data/ML_ready/participant_data.csv")

colnames(data_c)[4]="time"
colnames(data_i)[4]="time"
colnames(data_p)[4]="time"

data_c = data_c[,c(-23,-25,-27)]
data_i = data_i[,c(-23,-25,-27)]
data_p = data_p[,c(-23,-25,-27)]
```

## Difference between single signal features
Some additional features about coordination can be created - linear combination of single signal features, most efficient might be difference (patient-interviewer).
```{r}
#quantify the difference between the interviewer's and participant's gesturing systems - patient-interviewer
difference_features = data_p[,5:25]-data_i[,5:25]

#rename the columns to avoid duplicate column names
col_names = colnames(difference_features)
col_names = paste(col_names, "diff", sep="_")
colnames(difference_features)=col_names

data_c = cbind(data_c,difference_features)
```

```{r}
library(readr)
write_csv(data_c, "clean_data/ML_ready/coordination_data_enhanced.csv")
write_csv(data_i, "clean_data/ML_ready/interviewer_data_enhanced.csv")
write_csv(data_p, "clean_data/ML_ready/participant_data_enhanced.csv")
```
