==> 06_Single_models_&_evaluation.Rmd <==
---
title: "05 Machine Learning"
author: "Dominik Klepl"
date: "11/10/2018"
output: html_document
---

# Required libraries
```{r}
pacman::p_load("caret", "DALEX", "caretEnsemble", "ggthemes")
source("Scripts/modelling_fuctions.R")
```



# 1 Participant models
The list of algorithms trained: 
- Model averaged neural net ('avNNet')
- eXtreme Gradient Boosting ('xgbTree')
- Random Forest ('rf')
- Naive Bayes ('naive_bayes')
- Penalized Logistic Regression ('plr)

## 1.2 Model training
NB boruta selected all features

Load participant data and create index
```{r}
boruta_p = read.csv("clean_data/ML_data/Participant_boruta.csv")
boruta_p = caret_friendly(boruta_p)
index = boruta_p$index
```


```{r}
fitControl = trainControl(
  method = "repeatedcv",
  number = 10,
  repeats = 20,
  adaptive = list(min = 5, alpha = 0.05, method = "gls", complete = TRUE),
  classProbs = TRUE,
  summaryFunction = twoClassSummary,
  allowParallel = F,
  verboseIter = FALSE,
  search = "random",
  index = index,
  savePredictions = "final",
  sampling = "down"
)
```

### Boruta
```{r}
boruta_avNNet = train(diagnosis ~ ., data = boruta_p$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_avNNet, file="Models/Participant/boruta_avNNet.rds")
rm(boruta_avNNet)

boruta_xboost = train(diagnosis ~ ., data = boruta_p$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_xboost, file="Models/Participant/boruta_xboost.rds")
rm(boruta_xboost)

boruta_rf = train(diagnosis ~ ., data = boruta_p$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_rf, file="Models/Participant/boruta_rf.rds")
rm(boruta_rf)

boruta_bayes = train(diagnosis ~ ., data = boruta_p$train, 
                 method = 'naive_bayes', 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_bayes, file="Models/Participant/boruta_bayes.rds")
rm(boruta_bayes)

boruta_logistic = train(diagnosis ~ ., data = boruta_p$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_logistic, file="Models/Participant/boruta_logistic.rds")
rm(boruta_logistic)
rm(boruta_p)
```

### Featexp
```{r}
featexp_p = read.csv("cle
                     an_data/ML_data/Participant_featexp.csv")
featexp_p = caret_friendly(featexp_p)

featexp_avNNet = train(diagnosis ~ ., data = featexp_p$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_avNNet, file="Models/Participant/featexp_avNNet.rds")
rm(featexp_avNNet)

featexp_xboost = train(diagnosis ~ ., data = featexp_p$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_xboost, file="Models/Participant/featexp_xboost.rds")
rm(featexp_xboost)

featexp_rf = train(diagnosis ~ ., data = featexp_p$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_rf, file="Models/Participant/featexp_rf.rds")
rm(featexp_rf)

featexp_bayes = train(diagnosis ~ ., data = featexp_p$train, 
                 method = "naive_bayes", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_bayes, file="Models/Participant/featexp_bayes.rds")
rm(featexp_bayes)

featexp_logistic = train(diagnosis ~ ., data = featexp_p$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_logistic, file="Models/Participant/featexp_logistic.rds")
rm(featexp_logistic)
rm(featexp_p)
```

### Uncorrelated
```{r}
uncorrelated_p = read.csv("clean_data/ML_data/Participant_uncorrelated.csv")
uncorrelated_p = caret_friendly(uncorrelated_p)

uncorrelated_avNNet = train(diagnosis ~ ., data = uncorrelated_p$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_avNNet, file="Models/Participant/uncorrelated_avNNet.rds")
rm(uncorrelated_avNNet)

uncorrelated_xboost = train(diagnosis ~ ., data = uncorrelated_p$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_xboost, file="Models/Participant/uncorrelated_xboost.rds")
rm(uncorrelated_xboost)

uncorrelated_rf = train(diagnosis ~ ., data = uncorrelated_p$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_rf, file="Models/Participant/uncorrelated_rf.rds")
rm(uncorrelated_rf)

uncorrelated_bayes = train(diagnosis ~ ., data = uncorrelated_p$train, 
                 method = "naive_bayes", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_bayes, file="Models/Participant/uncorrelated_bayes.rds")
rm(uncorrelated_bayes)

uncorrelated_logistic = train(diagnosis ~ ., data = uncorrelated_p$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_logistic, file="Models/Participant/uncorrelated_logistic.rds")
rm(uncorrelated_logistic)
rm(uncorrelated_p)
```


# 2 Interviewer models

## 2.2 Model training

Load interviewer data and create index
```{r}
boruta_i = read.csv("clean_data/ML_data/Interviewer_boruta.csv")
boruta_i = caret_friendly(boruta_i)
index = boruta_i$index
```


```{r}
fitControl = trainControl(
  method = "repeatedcv",
  number = 10,
  repeats = 20,
  adaptive = list(min = 5, alpha = 0.05, method = "gls", complete = TRUE),
  classProbs = TRUE,
  summaryFunction = twoClassSummary,
  allowParallel = F,
  verboseIter = FALSE,
  search = "random",
  index = index,
  savePredictions = "final"
)
```

### Boruta
```{r}
boruta_avNNet = train(diagnosis ~ ., data = boruta_i$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_avNNet, file="Models/Interviewer/boruta_avNNet.rds")
rm(boruta_avNNet)

boruta_xboost = train(diagnosis ~ ., data = boruta_i$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_xboost, file="Models/Interviewer/boruta_xboost.rds")
rm(boruta_xboost)

boruta_rf = train(diagnosis ~ ., data = boruta_i$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_rf, file="Models/Interviewer/boruta_rf.rds")
rm(boruta_rf)

boruta_bayes = train(diagnosis ~ ., data = boruta_i$train, 
                 method = 'naive_bayes', 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_bayes, file="Models/Interviewer/boruta_bayes.rds")
rm(boruta_bayes)

boruta_logistic = train(diagnosis ~ ., data = boruta_i$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_logistic, file="Models/Interviewer/boruta_logistic.rds")
rm(boruta_logistic)
rm(boruta_i)
```

### Featexp
```{r}
featexp_i = read.csv("clean_data/ML_data/Interviewer_featexp.csv")
featexp_i = caret_friendly(featexp_i)

featexp_avNNet = train(diagnosis ~ ., data = featexp_i$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_avNNet, file="Models/Interviewer/featexp_avNNet.rds")
rm(featexp_avNNet)

featexp_xboost = train(diagnosis ~ ., data = featexp_i$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_xboost, file="Models/Interviewer/featexp_xboost.rds")
rm(featexp_xboost)

featexp_rf = train(diagnosis ~ ., data = featexp_i$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_rf, file="Models/Interviewer/featexp_rf.rds")
rm(featexp_rf)

featexp_bayes = train(diagnosis ~ ., data = featexp_i$train, 
                 method = "naive_bayes", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_bayes, file="Models/Interviewer/featexp_bayes.rds")
rm(featexp_bayes)

featexp_logistic = train(diagnosis ~ ., data = featexp_i$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_logistic, file="Models/Interviewer/featexp_logistic.rds")
rm(featexp_logistic)
rm(featexp_p)
```

### Uncorrelated
```{r}
uncorrelated_i = read.csv("clean_data/ML_data/Interviewer_uncorrelated.csv")
uncorrelated_i = caret_friendly(uncorrelated_i)

uncorrelated_avNNet = train(diagnosis ~ ., data = uncorrelated_i$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_avNNet, file="Models/Interviewer/uncorrelated_avNNet.rds")
rm(uncorrelated_avNNet)

uncorrelated_xboost = train(diagnosis ~ ., data = uncorrelated_i$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_xboost, file="Models/Interviewer/uncorrelated_xboost.rds")
rm(uncorrelated_xboost)

uncorrelated_rf = train(diagnosis ~ ., data = uncorrelated_i$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_rf, file="Models/Interviewer/uncorrelated_rf.rds")
rm(uncorrelated_rf)

uncorrelated_bayes = train(diagnosis ~ ., data = uncorrelated_i$train, 
                 method = "naive_bayes", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_bayes, file="Models/Interviewer/uncorrelated_bayes.rds")
rm(uncorrelated_bayes)

uncorrelated_logistic = train(diagnosis ~ ., data = uncorrelated_i$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_logistic, file="Models/Interviewer/uncorrelated_logistic.rds")
rm(uncorrelated_logistic)
rm(uncorrelated_i)
```


# 3. Coordination

Load data and create index
```{r}
boruta_c = read.csv("clean_data/ML_data/Coordination_boruta.csv")
boruta_c = caret_friendly(boruta_c)
index = boruta_c$index
```

```{r}
fitControl = trainControl(
  method = "repeatedcv",
  number = 10,
  repeats = 20,
  adaptive = list(min = 5, alpha = 0.05, method = "gls", complete = TRUE),
  classProbs = TRUE,
  summaryFunction = twoClassSummary,
  allowParallel = F,
  verboseIter = FALSE,
  search = "random",
  index = index,
  savePredictions = "final"
)
```

### Boruta
```{r}
boruta_avNNet = train(diagnosis ~ ., data = boruta_c$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_avNNet, file="Models/Coordination/boruta_avNNet.rds")
rm(boruta_avNNet)

boruta_xboost = train(diagnosis ~ ., data = boruta_c$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_xboost, file="Models/Coordination/boruta_xboost.rds")
rm(boruta_xboost)

boruta_rf = train(diagnosis ~ ., data = boruta_c$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_rf, file="Models/Coordination/boruta_rf.rds")
rm(boruta_rf)

boruta_bayes = train(diagnosis ~ ., data = boruta_c$train, 
                 method = 'naive_bayes', 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_bayes, file="Models/Coordination/boruta_bayes.rds")
rm(boruta_bayes)

boruta_logistic = train(diagnosis ~ ., data = boruta_c$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(boruta_logistic, file="Models/Coordination/boruta_logistic.rds")
rm(boruta_logistic)
rm(boruta_c)
```

### Featexp
```{r}
featexp_c = read.csv("clean_data/ML_data/Coordination_featexp.csv")
featexp_c = caret_friendly(featexp_c)

featexp_avNNet = train(diagnosis ~ ., data = featexp_c$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_avNNet, file="Models/Coordination/featexp_avNNet.rds")
rm(featexp_avNNet)

featexp_xboost = train(diagnosis ~ ., data = featexp_c$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_xboost, file="Models/Coordination/featexp_xboost.rds")
rm(featexp_xboost)

featexp_rf = train(diagnosis ~ ., data = featexp_c$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_rf, file="Models/Coordination/featexp_rf.rds")
rm(featexp_rf)

featexp_bayes = train(diagnosis ~ ., data = featexp_c$train, 
                 method = 'naive_bayes', 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_bayes, file="Models/Coordination/featexp_bayes.rds")
rm(featexp_bayes)

featexp_logistic = train(diagnosis ~ ., data = featexp_c$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(featexp_logistic, file="Models/Coordination/featexp_logistic.rds")
rm(featexp_logistic)
rm(featexp_c)
```

### Uncorrelated
```{r}
uncorrelated_c = read.csv("clean_data/ML_data/Coordination_uncorrelated.csv")
uncorrelated_c = caret_friendly(uncorrelated_c)

uncorrelated_avNNet = train(diagnosis ~ ., data = uncorrelated_c$train, 
                 method = "avNNet", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_avNNet, file="Models/Coordination/uncorrelated_avNNet.rds")
rm(uncorrelated_avNNet)

uncorrelated_xboost = train(diagnosis ~ ., data = uncorrelated_c$train, 
                 method = "xgbTree", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_xboost, file="Models/Coordination/uncorrelated_xboost.rds")
rm(uncorrelated_xboost)

uncorrelated_rf = train(diagnosis ~ ., data = uncorrelated_c$train, 
                 method = "rf", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_rf, file="Models/Coordination/uncorrelated_rf.rds")
rm(uncorrelated_rf)

uncorrelated_bayes = train(diagnosis ~ ., data = uncorrelated_c$train, 
                 method = 'naive_bayes', 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_bayes, file="Models/Coordination/uncorrelated_bayes.rds")
rm(uncorrelated_bayes)

uncorrelated_logistic = train(diagnosis ~ ., data = uncorrelated_c$train, 
                 method = "plr", 
                 trControl = fitControl,
                 metric = "ROC",
                 preProcess = c("center", "scale"))

saveRDS(uncorrelated_logistic, file="Models/Coordination/uncorrelated_logistic.rds")
rm(uncorrelated_logistic)
rm(uncorrelated_c)
```

# 4. Model Evaluation
```{r}
library(caret)
library(ggthemes)
```

## 4.1 Participant

Load all trained models
```{r}
models_participant = load_models(which = "P")
```

## Resamples analysis
GOAL - determine the feature set to use for fitting and tuning the final model
It's impossible to read anything out of the outputs/plots when all 12 models are examined.
=> Pick the best model from each feature-set first. Then compare the best models to determine which feature-set to use to consider for a meta-model.

```{r}
resamples_p = resamples(models_participant)
values = resamples_p$values[,(endsWith(names(resamples_p$values),"ROC"))]
mean = sapply(values, FUN=mean)
sd = sapply(values, FUN=sd)
best = which.max(mean)

mean = as.data.frame(mean)
sd = as.data.frame(sd)

roc_participant = cbind(mean,sd)
roc_participant$model = names(models_participant)
roc_participant$feature_set = as.factor(c(rep("boruta",5), rep("featexp",5), rep("uncorrelated",5)))
roc_participant = roc_participant[order(roc_participant$mean, decreasing = T),]
row.names(roc_participant) = NULL
roc_participant$rank = as.numeric(row.names(roc_participant))

library(tidyverse)
ranking_participant= roc_participant %>% group_by(feature_set) %>% summarise(sum=sum(rank))

library(ggthemes)
roc_p = ggplot(roc_participant, aes(x=reorder(model, -mean), y=mean))+
  geom_point()+
  geom_errorbar(aes(ymin=(mean+sd), ymax=(mean-sd)))+
  theme_few()+
  theme(axis.text.x = element_text(angle = 60, hjust = 1, size = 10, lineheight = 8))+
  labs(x="Participant", y="Area under the curve")
roc_p

ggsave("Model_evaluation/ROC_participant.jpg", roc_p, width = 7, height = 4)
write.csv(roc_participant, "Model_evaluation/ROC_participant.csv", row.names = F)
write.csv(ranking_participant, "Model_evaluation/Feature_set_ranking_participant.csv", row.names = F)
```


## 4.2 Coordination
Load all trained models
```{r}
models_coordination = load_models(which = "C")
```

## Resamples analysis
```{r}
resamples_c = resamples(models_coordination)
values = resamples_c$values[,(endsWith(names(resamples_c$values),"ROC"))]
mean = sapply(values, FUN=mean)
sd = sapply(values, FUN=sd)
mean = as.data.frame(mean)
sd = as.data.frame(sd)

roc_coordination = cbind(mean,sd)
roc_coordination$model = names(models_coordination)
roc_coordination$feature_set = as.factor(c(rep("boruta",5), rep("featexp",5), rep("uncorrelated",5)))
roc_coordination = roc_coordination[order(roc_coordination$mean, decreasing = T),]
row.names(roc_coordination) = NULL
roc_coordination$rank = as.numeric(row.names(roc_coordination))


ranking_coordination= roc_coordination %>% group_by(feature_set) %>% summarise(sum=sum(rank))

roc_c = ggplot(roc_coordination, aes(x=reorder(model, -mean), y=mean))+
  geom_point()+
  geom_errorbar(aes(ymin=(mean+sd), ymax=(mean-sd)))+
  theme_few()+
  theme(axis.text.x = element_text(angle = 60, hjust = 1, size = 10, lineheight = 8))+
  labs(x="Coordination", y=NULL)

ggsave("Model_evaluation/ROC_coordination.jpg", roc_c, width = 7, height = 4)
write.csv(roc_coordination, "Model_evaluation/ROC_coordination.csv", row.names = F)
write.csv(feat_set_ranked, "Model_evaluation/Feature_set_ranking_coordination.csv", row.names = F)

#plot roc_p and roc_c in one plot
library(gridExtra)
library(grid)

ROC_all = grid.arrange(roc_p, roc_c, clip="on", ncol=2, top=textGrob("AUC of all fitted models", gp=gpar(fontsize=18, font=3)))
ggsave("Model_evaluation/ROC_all_models.jpg", ROC_all, width = 8, height = 3)
```


## 4.3 Interviewer
Load all trained models
```{r}
models_interviewer = load_models(which = "I")
```

## Resamples analysis
```{r}
resamples_i = resamples(models_interviewer)
values = resamples_i$values[,(endsWith(names(resamples_i$values),"ROC"))]
mean = sapply(values, FUN=mean)
sd = sapply(values, FUN=sd)
mean = as.data.frame(mean)
sd = as.data.frame(sd)

roc_interviewer = cbind(mean,sd)
roc_interviewer$model = names(models_interviewer)
roc_interviewer$feature_set = as.factor(c(rep("boruta",5), rep("featexp",5), rep("uncorrelated",5)))
roc_interviewer = roc_interviewer[order(roc_interviewer$mean, decreasing = T),]
row.names(roc_interviewer) = NULL
roc_interviewer$rank = as.numeric(row.names(roc_interviewer))


ranking_interviewer= roc_interviewer %>% group_by(feature_set) %>% summarise(sum=sum(rank))

roc_i = ggplot(roc_interviewer, aes(x=reorder(model, -mean), y=mean))+
  geom_point()+
  geom_errorbar(aes(ymin=(mean+sd), ymax=(mean-sd)))+
  theme_few()+
  theme(axis.text.x = element_text(angle = 60, hjust = 1, size = 10, lineheight = 8))+
  labs(x="Interviewer", y=NULL)

ggsave("Model_evaluation/ROC_interviewer.jpg", roc_c, width = 7, height = 4)
write.csv(roc_interviewer, "Model_evaluation/ROC_interviewer.csv", row.names = F)
write.csv(feat_set_ranked, "Model_evaluation/Feature_set_ranking_interviewer.csv", row.names = F)

#plot roc_p and roc_c in one plot
library(gridExtra)
library(grid)

ROC_all = grid.arrange(roc_p, roc_c, clip="on", ncol=2, top=textGrob("AUC of all fitted models", gp=gpar(fontsize=18, font=3)))
ggsave("Model_evaluation/ROC_all_models.jpg", ROC_all, width = 8, height = 3)
```


#ENSEMBLE MODEL

Now we know which feature set to use for both participant and coordination models.
 => Participant - boruta set
 => Coordination - featexp set
Now we'll look at correlation between models predictions to find out if it even makes sense to build a meta-model (combination of predictions using glm).

```{r}
best_participant = models_participant[(startsWith(names(models_participant),"boruta"))]
best_coordination = models_coordination[(startsWith(names(models_coordination), "featexp"))]

#rename the models to correct names
names(best_participant) = c("avNNet", "NB", "PLR", "RF", "XGB")
names(best_coordination) = c("avNNet", "NB", "PLR", "RF", "XGB")

res_p = resamples(best_participant)
res_c = resamples(best_coordination)

correlation_p = modelCor(res_p)
correlation_c = modelCor(res_c)

#plot correlation matrix
library(corrplot)
corrplot(correlation_p, type = "upper", method = "color", addCoef.col = "black", tl.col = "black", tl.srt = 30)

corrplot(correlation_c, type = "upper", method = "color", addCoef.col = "black", tl.col = "black", tl.srt = 30)

#prediction correlation
p_data = caret_friendly(read.csv("clean_data/ML_data/Participant_boruta.csv"))
c_data = caret_friendly(read.csv("clean_data/ML_data/Coordination_featexp.csv"))

p_predictions = data.frame()
for (i in 1:length(best_participant)) {
  model = best_participant[[i]]
  name = model$method
  one_pred = as.data.frame(predict(model, p_data$X_test, type="prob")[,1])
  one_pred$model = name
  one_pred$n = 1:nrow(one_pred)
  p_predictions = rbind(p_predictions, one_pred)
}

c_predictions = data.frame()
for (i in 1:length(best_coordination)) {
  model = best_coordination[[i]]
  name = model$method
  one_pred = as.data.frame(predict(model, c_data$X_test, type="prob")[,1])
  one_pred$model = name
  one_pred$n = 1:nrow(one_pred)
  c_predictions = rbind(c_predictions, one_pred)
}

colnames(p_predictions)[1]="prob"
colnames(c_predictions)[1]="prob"

p_wide = data.frame(avNNet=p_predictions[p_predictions$model=="avNNet",1],
                    naive_bayes=p_predictions[p_predictions$model=="naive_bayes",1],
                    plr=p_predictions[p_predictions$model=="plr",1],
                    rf=p_predictions[p_predictions$model=="rf",1],
                    xgbTree=p_predictions[p_predictions$model=="xgbTree",1])
c_wide = data.frame(avNNet=c_predictions[c_predictions$model=="avNNet",1],
                    naive_bayes=c_predictions[c_predictions$model=="naive_bayes",1],
                    plr=c_predictions[c_predictions$model=="plr",1],
                    rf=c_predictions[c_predictions$model=="rf",1],
                    xgbTree=c_predictions[c_predictions$model=="xgbTree",1])

cor_matrix_p = cor(p_wide)
cor_matrix_c = cor(c_wide)


library(corrplot)
corrplot(cor_matrix_p, method = "color", addCoef.col = "black", tl.col = "black", tl.srt = 30)
corrplot(cor_matrix_c, method = "color", addCoef.col = "black", tl.col = "black", tl.srt = 30)
```

Both sets of model show high levels of correlation. Building an ensemble probably won't see a significant increase in performance.

Let's try it anyway
To train an ensemble we need to fit a model list: list of models fitted on the same CV folds in the same order.

Load data and CV indexes
```{r}
p_data = caret_friendly(read.csv("clean_data/ML_data/Participant_boruta.csv"))
c_data = caret_friendly(read.csv("clean_data/ML_data/Coordination_boruta.csv"))

index_p = readRDS("Models/Participant_index.rds")
index_c = readRDS("Models/Coordination_index.rds")
```

Fit the 2 model lists
```{r}
fitControl = trainControl(
  method = "repeatedcv",
  number = 10,
  repeats = 30,
  classProbs = TRUE,
  summaryFunction = twoClassSummary,
  savePredictions = "final",
  verboseIter = F,
  index = index_p
)

model_list_p <- caretList(
  x=p_data$X_train, y=p_data$Y_train,
  trControl=fitControl,
  methodList=c("gbm")
  )

model_list_p[["avNNet"]]=best_participant$avNNet
model_list_p[["naive_bayes"]]=best_participant$NB
model_list_p[["plr"]]=best_participant$PLR
model_list_p[["rf"]]=best_participant$RF
model_list_p[["xgbTree"]]=best_participant$XGB

fitControl = trainControl(
  method = "repeatedcv",
  number = 10,
  repeats = 30,
  classProbs = TRUE,
  summaryFunction = twoClassSummary,
  savePredictions = "final",
  verboseIter = F,
  index = index_c
)

model_list_c <- caretList(
  x=c_data$X_train, y=c_data$Y_train,
  trControl=fitControl,
  methodList=c("gbm")
  )

model_list_c[["avNNet"]]=best_coordination$avNNet
model_list_c[["naive_bayes"]]=best_coordination$NB
model_list_c[["plr"]]=best_coordination$PLR
model_list_c[["rf"]]=best_coordination$RF
model_list_c[["xgbTree"]]=best_coordination$XGB

saveRDS(model_list_p, "Models/Participant_list.rds")
saveRDS(model_list_c, "Models/Coordination_list.rds")
```


Train ensemble
```{r}
ensemble_p <- caretStack(
  model_list_p,
  method="glm",
  metric="ROC",
  trControl=trainControl(
    method="repeatedcv",
    number= 10,
    repeats = 10,
    savePredictions="final",
    classProbs=TRUE,
    summaryFunction=twoClassSummary)
  )

saveRDS(ensemble_p, "Models/Participant_ensemble")

ensemble_c <- caretStack(
  model_list_c,
  method="glm",
  metric="ROC",
  trControl=trainControl(
    method="repeatedcv",
    number= 10,
    repeats = 10,
    savePredictions="final",
    classProbs=TRUE,
    summaryFunction=twoClassSummary)
  )

saveRDS(ensemble_p, "Models/Participant_ensemble")
```



#CODE MESS for evaluation
```{r}
p_fun <- function(object, newdata){predict(object, newdata=newdata, type="prob")[,1]}
p_data$Y_test = ifelse(p_data$Y_test=="Control", 0, 1)
test = cbind(p_data$Y_test, p_data$X_test)
expl_b = explain(p_models$boruta_bayes, test[2:24], test[,1], predict_function = p_fun, label = "Naive Bayes")
expl_a = explain(p_models$boruta_avNNet, test[2:24], test[,1], predict_function = p_fun, label = "avNNet")

perf_b = model_performance(expl_b)
perf_a = model_performance(expl_a)

plot(perf_a, perf_b)

imp_a = variable_importance(expl_a, loss_function = loss_root_mean_square)
imp_b = variable_importance(expl_b, loss_function = loss_root_mean_square)

plot(imp_a, imp_b)
```
